React - Authentication Context - Movies and Profile both protected
In this application, the Movies and Profile routes are protected.
Click on the movies link you'll see that the Login Page is shown.
Click on the LogIn button and it will Authenticate you, allowing access to the protected groups.
Look at the code and notice that there is a pages.js that exports most of the pages for the app and the authContext.js provides the authentication context for all pages.